pyproject.toml
[build-system]
  build-backend = "uv_build"
  requires = ["uv_build"]


[dependency-groups]
  dev = ["dycw-utilities[test]>=0.189.1", "pyright>=1.1.408"]


[project]
  authors = [{ email = "d.wan@icloud.com", name = "Derek Wan" }]
  dependencies = ["click>=8.3.1", "dycw-utilities>=0.189.1"]
  description = "Library to operate `postgres` and `pgbackrest`"
  name = "dycw-postgres"
  readme = "README.md"
  requires-python = ">= 3.12"
  version = "0.1.2"


[tool]
  [tool.uv]
    [tool.uv.build-backend]
      module-name = "postgres"
      module-root = "src"
